main/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
python-dotenv
drf-spectacular
pandas
selenium
psycopg2-binary
python-decouple
dj-database-url